Fixed new game being played after winner found.
Fixed new game by resetting  global variables to default state.
New game is now playable even after a game has been played. Combined
code in  2 events files... they were too modular. Created a new variable
for each play so that API is updating each play correctly  instead of
swapping play then update first
