+++ assets/scripts/game/api.js
@@ -1,15 +1,16 @@
 'use strict'
 
-const config = require('../config.js')
-const store = require('../store.js')
+const config = require('./../config.js')
+const store = require('./../store.js')
 
-const create = function (data) {
+const create = function () {
   return $.ajax({
     url: config.apiUrl + '/games',
     method: 'POST',
     headers: {
       Authorization: 'Token token=' + store.user.token
-    }
+    },
+    data: '{}'
   })
 }
