+++ assets/scripts/game/ui.js
@@ -1,7 +1,6 @@
 'use strict'
 
-const store = require('../store.js')
-
+const store = require('./../store.js')
 
 const message = function (newText) {
   $('#game-message').text(newText)
@@ -30,10 +29,32 @@ const onWinningO = function () {
   message('Player O wins!')
 }
 
+const userMessage = function (newText) {
+  $('#game-message').text(newText)
+}
+const alertMessage = function (newText) {
+  $('#alert-message').text(newText)
+}
+
+const invalidMoveMessage = function () {
+  alertMessage('Invalid move! Pick another!')
+}
+
+const currentPlayerXMessage = function () {
+  userMessage(`Make your move player X `)
+}
+
+const currentPlayerOMessage = function () {
+  userMessage('Make your move player O')
+}
+
 module.exports = {
   onCreateNewGameSuccess,
   onUpdateGameSuccess,
   onCreateNewGameFailure,
   onWinningX,
-  onWinningO
+  onWinningO,
+  invalidMoveMessage,
+  currentPlayerXMessage,
+  currentPlayerOMessage
 }
